% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_exd.R
\name{assert_exd}
\alias{assert_exd}
\title{Assert that an experimental design has the correct columns}
\usage{
assert_exd(df)
}
\arguments{
\item{df}{data frame: experimental design for proteomics experiment}
}
\value{
\code{TRUE} if experimental design has correct columns.

Prints an error messages if there are extra and/or missing columns, with
the names of those extra and/or missing columns.
}
\description{
Certain plot functions adapted from the \href{doi.org/10.18129/B9.bioc.DEP}{DEP}
package require an experimental design to define data groups and labels.
This internal function ensures that only the exact, correct columns are
present in the experimental design data frame.
}
\details{
Extraneous columns are named so they can be removed and missing
columns are also named so they can be added to the experimental
design.
}
\examples{
# example of a correct experimental design
exp <- c(ID = c("A", "B", "C"),
         label = c("LFQ.intensity.A", "LFQ.intensity.B", "LFQ.intensity.C"),
         condition = c("sample", "sample", "control"),
         replicate = c(1, 2, 1))

}
\seealso{
This function is used in \code{\link{plot_heatmap2}}, etc.
}
