% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atcc.R
\docType{data}
\name{atcc}
\alias{atcc}
\title{Yeast (ATCC10231 and ATCC90028) LC-MS/MS data}
\format{A data frame with x rows and y variables:
\describe{
  \item{Protein.names}{Description}
  \item{Unique.peptides}{Unique peptides}
  ...
}}
\source{
Source goes here
}
\usage{
atcc
}
\description{
A description.
}
\keyword{datasets}
