% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atcc_exp.R
\docType{data}
\name{atcc_exp}
\alias{atcc_exp}
\title{Yeast (ATCC10231 and ATCC90028) LC-MS/MS experimental design}
\format{A data frame with x rows and y variables:
\describe{
  \item{ID}{Description}
  \item{label}{Description}
  \item{condition}{Description}
  \item{replicate}{Description}
}}
\source{
Made by hand
}
\usage{
atcc_exp
}
\description{
A description.
}
\keyword{datasets}
