% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biofilm_exp.R
\docType{data}
\name{biofilm_exp}
\alias{biofilm_exp}
\title{Biofilm LC-MS/MS experimental design}
\format{A data frame with x rows and y variables:
\describe{
  \item{ID}{Description}
  \item{label}{Description}
  \item{condition}{Description}
  \item{replicate}{Description}
}}
\source{
Made by hand
}
\usage{
biofilm_exp
}
\description{
A description.
}
\keyword{datasets}
