% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgd.R
\docType{data}
\name{cgd}
\alias{cgd}
\title{Candida Genome Database reference table}
\format{A data frame with x rows and y variables:
\describe{
  \item{Systematic_name}{Description}
  \item{Feature_name}{Description}
  ...
}}
\source{
\url{http://www.candidagenome.org/download/chromosomal_feature_files/C_albicans_SC5314/}{Downloaded date}
}
\usage{
cgd
}
\description{
A description.
}
\keyword{datasets}
