% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_result.R
\name{combine_result}
\alias{combine_result}
\title{Title}
\usage{
combine_result(mat, toptab)
}
\arguments{
\item{mat}{Description}

\item{toptab}{Description}
}
\value{
Description
}
\description{
Title
}
\examples{
# example
}
