% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunns_test.R
\name{dunns_test}
\alias{dunns_test}
\title{Title}
\usage{
dunns_test(x, g, method, ...)
}
\arguments{
\item{x}{Description.}

\item{g}{Description.}

\item{method}{Description.}

\item{...}{Description.}
}
\value{
Description.
}
\description{
Description.
}
\examples{
# example

}
