% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_blank.R
\name{fill_blank}
\alias{fill_blank}
\title{Fill blanks in character column}
\usage{
fill_blank(df, blank_col, fill_col)
}
\arguments{
\item{df}{Data frame}

\item{blank_col}{Data frame character column containing empty strings}

\item{fill_col}{Data frame column containing strings to replace empty strings}
}
\value{
\code{df} with empty strings in a specific character column filled
with strings from another column
}
\description{
\code{fill_blank} fills any empty "" strings in a data frame character
column with the corresponding strings from another character column in
the same data frame.
}
\examples{
# make data frame with blanks
# columns must be class character
df <- data.frame(name1 = c("Bob", "Jane", "Henry", "Sue"),
                 name2 = c("Mary", "", "Tim", ""),
                 stringsAsFactors = FALSE)

# fill blanks in names2
df <- fill_blank(df, "name2", "name1")

}
