% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_na.R
\name{filter_na}
\alias{filter_na}
\title{Title}
\usage{
filter_na(data, op = c("==", "<=", ">="), pat, val)
}
\arguments{
\item{data}{Description}

\item{op}{Description}

\item{pat}{Description}

\item{val}{Description}
}
\value{
Returns x
}
\description{
Title
}
\examples{
# example

}
