% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_na2.R
\name{filter_na2}
\alias{filter_na2}
\title{Title}
\usage{
filter_na2(data, logic = c("and", "or"), op = c("==", ">=", "<="),
  pat1, val1, pat2, val2)
}
\arguments{
\item{data}{Description}

\item{logic}{Description}

\item{op}{Description}

\item{pat1}{Description}

\item{val1}{Description}

\item{pat2}{Description}

\item{val2}{Description}
}
\value{
Returns x
}
\description{
Title
}
\examples{
# example

}
