% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_val4.R
\name{filter_val4}
\alias{filter_val4}
\title{Title}
\usage{
filter_val4(data, logic = c("and", "or"), op = c("==", ">=", "<="),
  pat1, val1, pat2, val2, pat3, val3, pat4, val4)
}
\arguments{
\item{data}{Description.}

\item{logic}{Description.}

\item{op}{Description.}

\item{pat1}{Description.}

\item{val1}{Description.}

\item{pat2}{Description.}

\item{val2}{Description.}

\item{pat3}{Description.}

\item{val3}{Description.}

\item{pat4}{Description.}

\item{val4}{Description.}
}
\value{
Description.
}
\description{
Description.
}
\examples{
# example

}
