% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_zero.R
\name{filter_zero}
\alias{filter_zero}
\title{Title}
\usage{
filter_zero(data, op = c("==", "<=", ">="), pat, val)
}
\arguments{
\item{data}{Description.}

\item{op}{Description.}

\item{pat}{Description.}

\item{val}{Description.}
}
\value{
Description.
}
\description{
Description.
}
\examples{
# example

}
