% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotation2.R
\name{get_annotation2}
\alias{get_annotation2}
\title{Title}
\usage{
get_annotation2(mat, exd, indicate)
}
\arguments{
\item{mat}{Description}

\item{exd}{Description}

\item{indicate}{Description}
}
\value{
Description
}
\description{
Internal function to get ComplexHeatmap::HeatmapAnnotation object.
}
\examples{
# example

}
