% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_results}
\alias{get_results}
\title{Title}
\usage{
get_results(efit, mat, p_val = 0.01, lfc = 0, type = c("overall",
  "individual"))
}
\arguments{
\item{efit}{Description}

\item{mat}{Description}

\item{p_val}{Description}

\item{lfc}{Description}

\item{type}{Description}
}
\value{
Description
}
\description{
Description.
}
\examples{
# example

}
