% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gilbona.R
\docType{data}
\name{gilbona}
\alias{gilbona}
\title{Gil-Bona LC-MS/MS data}
\format{A data frame with x rows and y variables:
\describe{
  \item{Feature_name}{Description}
  \item{Gene_name}{Description}
  ...
}}
\source{
\url{doi.org/10.1021/pr5007944}{Downloaded date}
}
\usage{
gilbona
}
\description{
A description.
}
\keyword{datasets}
