% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_eBayes.R
\name{limma_eBayes}
\alias{limma_eBayes}
\title{Title}
\usage{
limma_eBayes(mat, design, contrasts)
}
\arguments{
\item{mat}{Description.}

\item{design}{Description.}

\item{contrasts}{Description.}
}
\value{
Description.
}
\description{
Description.
}
\examples{
#example

}
