% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_id.R
\name{match_id}
\alias{match_id}
\title{Replace values based on match}
\usage{
match_id(id, ref, match, new)
}
\arguments{
\item{id}{character vector: the values to be matched}

\item{ref}{data frame: reference table with values to be matched against}

\item{match}{string: column name of character column in reference table with
values to be matched against}

\item{new}{string: column name of character column in reference table with new
information to replace matches}
}
\value{
A vector with the same length as \code{id} but with the values replaced
with corresponding information from the reference table.

If there is a value to be matched which is not present in the reference table,
the value will be replaced with the string "NA"
}
\description{
\code{match_id} matches a character vector with values (which may or may not be
separated by semicolons) to a reference table with a character column containing
values (which must not be separated by semicolons) and then replaces those
values with corresponding information from another character column in the
reference table.
}
\examples{
# load dplyr
library(dplyr)

# make a reference table
ref_df <- data.frame(id = c("a1", "a2", "a3"), last_name = c("Smith", "Brown", "Doe"))

# make a data frame
my_names <- data.frame(id = c("a1;a3", "a10", "a2"), first = c("Jane", "Mary", "John"))

# use match_id to make new column in my_names
my_names$last <- match_id(my_names$id, ref_df, "id", "last_name")

}
