% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_id_multi.R
\name{match_id_multi}
\alias{match_id_multi}
\title{Title}
\usage{
match_id_multi(df, id, ref, match, new, as_list = FALSE)
}
\arguments{
\item{df}{Description}

\item{id}{Description}

\item{ref}{Description}

\item{match}{Description}

\item{new}{Description}

\item{as_list}{Description}
}
\value{
Description
}
\description{
Description.
}
\examples{
# make a reference table
ref_df <- data.frame(id = c("a1", "a2", "a3"),
                     last_name = c("Smith", "Brown", "Doe"),
                     atr1 = c("Blue", "Brown", "Yellow"),
                     atr2 = c("Big", "Medium", "Small"),
                     stringsAsFactors = FALSE)

# make a dataset to which you want to append columns
# note that row 3 has an id which is NOT in the reference table
my_names <- data.frame(employee_id = c("a1;a3", "a2", "a5"),
                       first_name = c("Jane", "Mary", "John"),
                       stringsAsFactors = FALSE)

# append selected columns to my_names based on matching ids
# note that row 3 has NA in all appended columns because
# the employee_id = a5 is not in the reference table
my_names <- match_id_multi(df = my_names,
                           id = "employee_id",
                           ref = ref_df,
                           match = "id",
                           new = c("last_name", "atr2", "atr1"))

# alternatively, return a list of character vectors for the
columns selected in by the 'new' argument
my_list <- match_id_multi(df = my_names,
                          id = "employee_id",
                          ref = ref_df,
                          match = "id",
                          new = c("atr1", "atr2"))

}
