% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nta.R
\docType{data}
\name{nta}
\alias{nta}
\title{\emph{C. albicans} nanoparticle tracking analysis (NTA) data}
\format{An object of class \code{data.frame} with 42,000 observations (rows) and 7 variables (columns).}
\source{
Dawson, C. S., Bleackley, M. R., Garcia-Ceron, D., & Anderson, M. A. (2019).
Title. \emph{Journal vol}(issue), 000-000. doi:
}
\usage{
nta
}
\description{
\code{nta} is a \code{data.frame} of particle size
count data measured via NTA for
extracellular vesicles (EVs) isolated from several different
\emph{C. albicans} strains. These data were generated by
Dawson et al. (2019) using the \code{\link[tidyNano]{nanocombine}} and
\code{\link[tidyNano]{nanotidy}} function
from the package \pkg{tidyNano}.

There are 14 samples total in this data set:
\itemize{
\item DAY286 yeast EV biological replicates 1, 2, 3
\item ATCC90028 yeast EV biological replicates 1, 2, 3
\item ATCC10231 yeast EV biological replicates 1, 2, 3
\item DAY286 biofilm EV biological replicates 1, 2, 3, 4, 5
}

Each of these samples was measured 3 times on the NTA, therefore there are
3 technical replicates per sample.

See \href{https://doi.org}{Dawson et al. (2019)} for more information about the
generation of this data set. The raw NTA output \code{.csv} files and the
script used to generate this data.frame can be found
\href{https://github.com/csdaw/candidaev/tree/master/data-raw/nta}{here}.
}
\details{
Each row represents 1 particle size 'bin' as measured by a Malvern Instruments
NS300 using NTA software version 3.2.

Each column is as follows:

\describe{
\item{\strong{particle_size}}{Centre of the particle size 'bin'.}
\item{\strong{Strain}}{\emph{C. albicans} strain from which the EVs were isolated.}
\item{\strong{Bio_rep}}{EV biological replicate.}
\item{\strong{Dilution}}{Dilution factor for EVs in PBS, performed prior to NTA measurements.}
\item{\strong{Tech_rep}}{NTA measurement technical replicate.}
\item{\strong{Count}}{Concentration of particles measured in that particular 'bin' (particles / mL).}
\item{\strong{True_count}}{Count adjusted for the EV dilution factor (Count * Dilution).}
}
}
\keyword{datasets}
