% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dendro.R
\name{plot_dendro}
\alias{plot_dendro}
\title{Plot dendrogram}
\usage{
plot_dendro(mat)
}
\arguments{
\item{mat}{numeric matrix or data.frame: \code{NA} are allowed}
}
\value{
Returns a dendrogram (generated by
\code{\link[ggdendro]{ggdendrogram}}).
}
\description{
\code{plot_dendro} generates a dendrogram based on heirarchical clustering
using Gower's formula via the \code{\link[cluster]{daisy}} function from the
package \pkg{cluster}.

This function allows for the presence of \code{NA} in the data set. According
to the \code{\link{daisy}} documentation, "missing values in a row of x
are not included in the dissimilarities involving that row."
}
\examples{
# make a numeric matrix
mat1 <- matrix(c(1,12,4, 2,13,3), nrow = 2, ncol = 3, byrow = TRUE,
dimnames = list(c("row1", "row2"), c("C.1", "C.2", "C.3")))

# plot a dendrogram
plot_dendro(mat1)

}
