% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_detect2.R
\name{plot_detect2}
\alias{plot_detect2}
\title{Visualise intensities of proteins with missing values}
\usage{
plot_detect2(mat)
}
\arguments{
\item{mat}{numeric matrix: must have missing values}
}
\value{
Returns density and cumulative sum plots of protein intensities for
proteins with and without missing values
(generated by \code{\link[ggplot2]{ggplot}}).
}
\description{
\code{plot_detect2} generates density and cumulative sum plots
of protein intensities for proteins with and without missing values.

This function has been taken directly from the \code{plot_detect} function
in the \href{http://doi.org/10.18129/B9.bioc.DEP}{DEP} package and then
modified for use in this package.

\href{http://doi.org/10.18129/B9.bioc.DEP}{DEP} is distributed with the Artistic-2.0 license
which requires any source code modifications to be clearly stated.

\code{plot_detect2} includes the following modifications of \code{plot_detect}:

\itemize{
\item Dependance on a "SummarizedExperiment" object as input has been removed.
Instead the input has been changed to a numeric matrix.
\item Use of \code{theme_DEP1()} has been replaced with \code{theme_bw()}
\item x axis labels have been changed to "log2 LFQ intensity".
}
}
\examples{
# load dplyr
library(dplyr)

# load a proteinGroups data.frame supplied with this package
my_proteinGroups <- atcc

# load its corresponding experimental design
my_expDesign <- atcc_exp

# filter for proteins identified with minimum 3 unique peptides
# and convert to numeric matrix
my_lfq <- my_proteinGroups \%>\%
  filter(Unique.peptides >= 3) \%>\%
  convert_lfq(., my_expDesign)

# filter for proteins quantified in min 2/3 reps of
# at least 1 sample group
my_filt <- my_lfq \%>\%
  filter_na4(., logic = "or", op = "<=",
             pat1 = "A10231_EV", val1 = 1,
             pat2 = "A10231_W", val2 = 1,
             pat3 = "A90028_EV", val3 = 1,
             pat4 = "A90028_W", val4 = 1)

# normalise LFQ intensities
my_norm <- limma::normalizeCyclicLoess(my_filt)

# generate density and cumulative sum plots
plot_detect2(my_norm)

}
\references{
Zhang, X., Smits, A., van Tilburg, G., Ovaa, H., Huber, W., Vermeulen,
M. (2018). Proteome-wide identification of ubiquitin interactions using UbIA-MS.
\emph{Nature Protocols}, \strong{13}, 530-550.
\href{https://doi.org/10.1038/nprot.2017.147}{doi:10.1038/nprot.2017.147}
}
