% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_frequency2.R
\name{plot_frequency2}
\alias{plot_frequency2}
\title{Plot frequency of protein quantification between samples}
\usage{
plot_frequency2(mat, plot = TRUE)
}
\arguments{
\item{mat}{numeric matrix: protein log2 LFQ intensity data}

\item{plot}{logical:
if \code{TRUE} (default) the barplot is produced.
Otherwise if \code{FALSE}, the data on which the
barplot is based are returned}
}
\value{
Returns a bar plot of the frequency of protein quantification
between samples
(generated by \code{\link[ggplot2]{ggplot}}). If \code{plot = FALSE},
returns a \code{data.frame} with the data underlying the bar plot.
}
\description{
\code{plot_frequency2} generates a bar plot of how many proteins are
quantified in only 1 out of x samples, 2 out of x samples, 3 out of
x samples, etc., i.e. protein quantification frequency.

This function has been taken directly from the \code{plot_frequency} function
in the \href{http://doi.org/10.18129/B9.bioc.DEP}{DEP} package and then
modified for use in this package.

\href{http://doi.org/10.18129/B9.bioc.DEP}{DEP} is distributed with the Artistic-2.0 license
which requires any source code modifications to be clearly stated.

\code{plot_frequency2} includes the following modifications of \code{plot_frequency}:

\itemize{
\item Dependance on a "SummarizedExperiment" object as input has been removed.
Instead the input has been changed to a numeric matrix.
\item Use of \code{theme_DEP1()} has been replaced with \code{theme_bw()}
\item Plot title has been changed to "Protein quantification frequency".
}
}
\examples{
# load dplyr
library(dplyr)

# load a proteinGroups data.frame supplied with this package
my_proteinGroups <- atcc

# load its corresponding experimental design
my_expDesign <- atcc_exp

# filter for proteins identified with minimum 3 unique peptides
# and convert to numeric matrix
my_lfq <- my_proteinGroups \%>\%
  filter(Unique.peptides >= 3) \%>\%
  convert_lfq(., my_expDesign)

# generate bar plot of protein quantification frequency
plot_frequency2(my_lfq)

}
