% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_go_bar.R
\name{plot_go_bar}
\alias{plot_go_bar}
\title{Title}
\usage{
plot_go_bar(go_table, terms_list, col_pal)
}
\arguments{
\item{go_table}{Description.}

\item{terms_list}{Description.}

\item{col_pal}{Description.}
}
\value{
Description.
}
\description{
Description.
}
\examples{
# example

}
