% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Title}
\usage{
plot_heatmap(mt, plot = TRUE, df = FALSE, data_type = c("log2fc",
  "log2intensity"), clust_fun = c("gower", "euclidean", "maximum",
  "manhattan", "canberra", "binary", "minkowski", "pearson", "spearman",
  "kendall"), split_type = c("kmeans", "cutree"), k = 5,
  cluster_split = TRUE, split_order = seq(1, k, 1),
  colour_lims = c(-8, 8), use_cbrewer = FALSE, cbrewer_pal = NULL,
  colour_vals = c("blue", "white", "red"), legend_pos = "top",
  row_title_fontsize = 6, col_name_fontsize = 10, ...)
}
\arguments{
\item{mt}{Description.}

\item{plot}{Description.}

\item{df}{Description.}

\item{data_type}{Description.}

\item{clust_fun}{Description.}

\item{split_type}{Description.}

\item{k}{Description.}

\item{cluster_split}{Description.}

\item{split_order}{Description.}

\item{colour_lims}{Description.}

\item{use_cbrewer}{Description.}

\item{cbrewer_pal}{Description.}

\item{colour_vals}{Description.}

\item{legend_pos}{Description.}

\item{row_title_fontsize}{Description.}

\item{col_name_fontsize}{Description.}

\item{...}{Description.}
}
\value{
Description.
}
\description{
Description.
}
\examples{

# example

}
