% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missval2.R
\name{plot_missval2}
\alias{plot_missval2}
\title{Plot a heatmap of proteins with missing values}
\usage{
plot_missval2(mat)
}
\arguments{
\item{mat}{numeric matrix: must have missing values}
}
\value{
Returns a heatmap indicating whether values are missing (white) or
not (black)
(generated by \code{\link[ComplexHeatmap]{Heatmap}}).
}
\description{
\code{plot_missval2} generates a heatmap of proteins
with missing values to observe the pattern of missing values across the
samples (random or not random).

This function has been taken directly from the \code{plot_missval}
function in the \href{http://doi.org/10.18129/B9.bioc.DEP}{DEP} package and
then modified for use in this package.

\href{http://doi.org/10.18129/B9.bioc.DEP}{DEP} is distributed with the
Artistic-2.0 license
which requires any source code modifications to be clearly stated.

\code{plot_missval2} includes the following modifications
of \code{plot_missval}:

\itemize{
\item Dependance on a "SummarizedExperiment" object as input has been
removed. Instead the input has been changed to a numeric matrix.
}
}
\examples{
# load dplyr
library(dplyr)

# load a proteinGroups data.frame supplied with this package
my_proteinGroups <- atcc

# load its corresponding experimental design
my_expDesign <- atcc_exp

# filter for proteins identified with minimum 3 unique peptides
# and convert to numeric matrix
my_lfq <- my_proteinGroups \%>\%
  filter(Unique.peptides >= 3) \%>\%
  convert_lfq(., my_expDesign)

# filter for proteins quantified in min 2/3 reps of
# at least 1 sample group
my_filt <- my_lfq \%>\%
  filter_na4(., logic = "or", op = "<=",
             pat1 = "A10231_EV", val1 = 1,
             pat2 = "A10231_W", val2 = 1,
             pat3 = "A90028_EV", val3 = 1,
             pat4 = "A90028_W", val4 = 1)

# plot missing value heatmap
plot_missval2(my_filt)

}
\references{
Zhang, X., Smits, A., van Tilburg, G., Ovaa, H., Huber, W., Vermeulen,
M. (2018). Proteome-wide identification of ubiquitin interactions using UbIA-MS.
\emph{Nature Protocols}, \strong{13}, 530-550.
\href{https://doi.org/10.1038/nprot.2017.147}{doi:10.1038/nprot.2017.147}
}
