% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_normalization2.R
\name{plot_normalization2}
\alias{plot_normalization2}
\title{Visualise effect of normalisation on protein intensity distribution}
\usage{
plot_normalization2(exd, facet_labels = NULL, mat, ...)
}
\arguments{
\item{exd}{data.frame: contains information for plot labels. See the
examples section of \code{\link{assert_exd}} or \code{\link{atcc_exp}} for
examples of appropriately structured experimental designs.}

\item{facet_labels}{\code{NULL} or a named vector: if\code{NULL} then
the matrix names will be used for the facet labels (aka. strip text). Use a
named vector to customise the facet labels. For example if you are plotting
the following: \code{plot_normalization2(exp, facet_labels = x, mat1, mat2)}
then \code{x} should be \code{c(mat1 = "Matrix 1", mat2 = "Matrix 2")}.}

\item{mat}{numeric matrix: LFQ intensity data before normalisation}

\item{...}{numeric matrix/matrices: LFQ intensity data after normalisation.
Multiple matrices can be input for plotting.}
}
\value{
Returns box plots of all input matrices, e.g. before and
after normalisation (generated by \code{\link[ggplot2]{ggplot}}).
}
\description{
\code{plot_normalization2} generates boxplots
of all conditions for input objects, e.g. before and after normalization.

This function has been taken directly from the \code{plot_normalization}
function in the \href{http://doi.org/10.18129/B9.bioc.DEP}{DEP} package and
then modified for use in this package.

\href{http://doi.org/10.18129/B9.bioc.DEP}{DEP} is distributed with the
Artistic-2.0 license
which requires any source code modifications to be clearly stated.

\code{plot_normalization2} includes the following modifications
of \code{plot_normalization}:

\itemize{
\item Dependance on a "SummarizedExperiment" object as input has been
removed. Instead the input has been changed to a numeric matrix.
\item Use of \code{theme_DEP1()} has been replaced with \code{theme_bw()}
}
}
\examples{
# load dplyr
library(dplyr)

# load a proteinGroups data.frame supplied with this package
my_proteinGroups <- atcc

# load its corresponding experimental design
my_expDesign <- atcc_exp

# filter for proteins identified with minimum 3 unique peptides
# and convert to numeric matrix
my_lfq <- my_proteinGroups \%>\%
  filter(Unique.peptides >= 3) \%>\%
  convert_lfq(., my_expDesign)

# filter for proteins quantified in min 2/3 reps of
# at least 1 sample group
my_filt <- my_lfq \%>\%
  filter_na4(., logic = "or", op = "<=",
             pat1 = "A10231_EV", val1 = 1,
             pat2 = "A10231_W", val2 = 1,
             pat3 = "A90028_EV", val3 = 1,
             pat4 = "A90028_W", val4 = 1)

# normalise LFQ intensities using limma
my_norm <- limma::normalizeCyclicLoess(my_filt)

# normalise LFQ intensities
# using some other function from limma
my_norm2 <- limma::normalizeQuantiles(my_filt)

# box plots before and after normalisations
plot_normalization2(exd = atcc_exp,
                    my_filt,
                    my_norm,
                    my_norm2)

}
