% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_numbers2.R
\name{plot_numbers2}
\alias{plot_numbers2}
\title{Plot protein numbers for each sample}
\usage{
plot_numbers2(mat, exd, plot = TRUE)
}
\arguments{
\item{mat}{numeric matrix: protein log2 LFQ intensity data}

\item{exd}{data.frame: contains information for plot labels. See the
examples section of \code{\link{assert_exd}} or \code{\link{atcc_exp}} for
examples of appropriately structured experimental designs.}

\item{plot}{logical: if \code{TRUE}, the bar plot will be drawn using
\code{\link[ggplot2]{ggplot}}. If \code{FALSE}, the data.frame underlying
the bar plot will be returned.}
}
\value{
Returns a barplot of the number of quantified proteins per sample
(generated by \code{\link[ggplot2]{ggplot}}). If \code{plot = FALSE},
returns a \code{data.frame} with the data underlying the bar plot.
}
\description{
\code{plot_numbers2} generates a bar plot
of the number of quantified proteins per sample.

This function has been taken directly from the \code{plot_numbers}
function in the \href{http://doi.org/10.18129/B9.bioc.DEP}{DEP} package and
then modified for use in this package.

\href{http://doi.org/10.18129/B9.bioc.DEP}{DEP} is distributed with the
Artistic-2.0 license
which requires any source code modifications to be clearly stated.

\code{plot_numbers2} includes the following modifications
of \code{plot_numbers}:

\itemize{
\item Dependance on a "SummarizedExperiment" object as input has been
removed. Instead the input has been changed to a numeric matrix.
\item Use of \code{theme_DEP1()} has been replaced with \code{theme_bw()}
}
}
\examples{
# load a proteinGroups data.frame supplied with this package
my_proteinGroups <- atcc

# load its corresponding experimental design
my_expDesign <- atcc_exp

# filter for proteins identified with minimum 3 unique peptides
# and convert to numeric matrix
my_lfq <- my_proteinGroups \%>\%
  filter(Unique.peptides >= 3) \%>\%
  convert_lfq(., my_expDesign)

# generate bar plot of number of proteins
# quantified per sample
plot_numbers2(my_lfq)

}
