% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_venn.R
\name{plot_venn}
\alias{plot_venn}
\title{Title}
\usage{
plot_venn(vlist, use_uniprot = FALSE, type = c("list", "plot", "save"),
  output = c("tiff", "png"), output_file = NULL, output_dim = c(1000,
  1000), output_res = 300, output_units = "px", output_pts = 14, ...)
}
\arguments{
\item{vlist}{Description}

\item{use_uniprot}{Description}

\item{type}{Description}

\item{output}{Description}

\item{output_dim}{Description}

\item{output_res}{Description}

\item{output_units}{Description}

\item{output_pts}{Description}

\item{...}{Description}
}
\value{
Description
}
\description{
Title
}
\examples{
# make two data frames
v1 <- data.frame(colour = c("green", "red", "blue"), level = c(60, 50, 45))

v2 <- data.frame(colour = c("purple", "green", "pink"), level = c(10, 13, 22))

# make list of vectors to compare
comp <- list(c1 = v1$colour, c2 = v2$colour)


# plot venn diagram
plot_venn(vlist = comp)

# plot venn diagram and save as png
plot_venn(vlist = comp, file = "my_venn.png", output = "png")

# don't plot venn but get venn partitions
part <- plot_venn(vlist = comp, plot = FALSE)

}
