% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_fungifun.R
\name{process_fungifun}
\alias{process_fungifun}
\title{Process FungiFun2 GO output}
\usage{
process_fungifun(filename)
}
\arguments{
\item{filename}{character: path to the .csv file output by FungiFun2}
}
\value{
Returns a data.frame with the results of the FungiFun2 GO term
enrichment analysis.
}
\description{
\code{process_fungifun} is a function that loads the
\code{.csv} output file from the online tool
\href{https://sbi.hki-jena.de/fungifun/}{FungiFun2} and cleans up the
table for analysis.

The settings that should be used for the FungiFun2 functional enrichment
analysis are:
\itemize{
\item Classification ontology = GO
\item 9/9 columns in the summary table should be shown.
\item Save as = CSV
}
}
\examples{
\donttest{
# write some gene names to the clipboard
# only works with Windows
my_genes <- c("GSC1", "XOG1", "BGL2", "HYR1", "ALS3",
              "PLB4.5", "SAP9")

writeClipboard(my_genes)

# paste into FungiFun2 input
# perform GO analysis and save output as CSV
# import GO enrichment analysis results
GO_result <- process_fungifun("./vignettes/figures/f4_GO_DAY_Y.csv")
}

}
