% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_dotdot.R
\name{remove_dotdot}
\alias{remove_dotdot}
\title{Remove duplicate . in column names}
\usage{
remove_dotdot(df)
}
\arguments{
\item{df}{data frame: column names contain duplicate .. or ... etc.}
}
\value{
data frame: column names contain only single .
}
\description{
Internal function to remove duplicated periods in column names with a
single period.
}
