% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_table.R
\name{sum_table}
\alias{sum_table}
\title{Summarise data table}
\usage{
sum_table(data)
}
\arguments{
\item{data}{data.frame or matrix: must only contain numeric data.}
}
\value{
Returns a tibble with the following columns:
\itemize{
\item \code{label}, name of the column being summarised
\item \code{mean}
\item \code{median}
\item \code{std.dev}, standard deviation
\item \code{coeff.var}, coefficient of variation
\item \code{min}
\item \code{max}
\item \code{range}
\item \code{n.valid.val}, number of non-\code{NA} values
\item \code{n.na.val}, number of \code{NA} values
\item \code{prcnt.na}, percentage of total values that are \code{NA}
}
}
\description{
\code{sum_table} generates a summary table (a tibble) of
descriptive statistics for each column of the input data table,
which can be a matrix or a data.frame.
}
\examples{
# make a matrix
my_mat <- matrix(c(2, 7, NA, NA,
                   3, 6, NA, NA,
                   1, NA, 8, NA,
                   432, 564, 789, 555),
                   nrow = 4,
                   ncol = 4)
colnames(my_mat) <- c("sample....1",
                      "sample..2",
                      "sample...3",
                      "total")

# summarise the matrix
sum_table(my_mat)

}
