% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukey_hsd.R
\name{tukey_hsd}
\alias{tukey_hsd}
\title{Title}
\usage{
tukey_hsd(x, ...)
}
\arguments{
\item{x}{Description.}

\item{...}{Description.}
}
\value{
Description.
}
\description{
Description.
}
\examples{
# example

}
