% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\docType{data}
\name{uniprot}
\alias{uniprot}
\title{UniProt Candida reference table}
\format{A data frame with x rows and y variables:
\describe{
  \item{CGD_gene_name}{Description}
  \item{UP_accession}{Description}
  ...
}}
\source{
\url{https://www.uniprot.org/proteomes/UP000000559}{Downloaded date}
}
\usage{
uniprot
}
\description{
A description.
}
\keyword{datasets}
