% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vargas.R
\docType{data}
\name{vargas}
\alias{vargas}
\title{Vargas LC-MS/MS data}
\format{A data frame with x rows and y variables:
\describe{
  \item{UP_accession}{Description}
  \item{var_gene_name}{Description}
  ...
}}
\source{
\url{doi.org/10.1111/cmi.12374}{Downloaded date}
}
\usage{
vargas
}
\description{
A description.
}
\keyword{datasets}
