% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wolf.R
\docType{data}
\name{wolf}
\alias{wolf}
\title{Wolf LC-MS/MS data}
\format{A data frame with x rows and y variables:
\describe{
  \item{UP_accession}{Description}
  \item{Gene_name}{Description}
  ...
}}
\source{
\url{doi.org/10.1128/EC.00054-15}{Downloaded date}
}
\usage{
wolf
}
\description{
A description.
}
\keyword{datasets}
