% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeast.R
\docType{data}
\name{yeast}
\alias{yeast}
\title{Yeast LC-MS/MS data}
\format{A data frame with x rows and y variables:
\describe{
  \item{Protein.names}{Description}
  \item{Unique.peptides}{Unique peptides}
  ...
}}
\source{
Source goes here
}
\usage{
yeast
}
\description{
A description.
}
\keyword{datasets}
