% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zarnowski.R
\docType{data}
\name{zarnowski}
\alias{zarnowski}
\title{Zarnowski LC-MS/MS data}
\format{A data frame with x rows and y variables:
\describe{
  \item{UP_accession}{Description}
  \item{Gene_name}{Description}
  ...
}}
\source{
\url{doi.org/10.1371/journal.pbio.2006872}{Downloaded date}
}
\usage{
zarnowski
}
\description{
A description.
}
\keyword{datasets}
