% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biofilm.R
\docType{data}
\name{biofilm}
\alias{biofilm}
\title{Biofilm LC-MS/MS data}
\format{A data frame with x rows and y variables:
\describe{
  \item{Majority.protein.IDs}{Description}
  \item{Unique.peptides}{Unique peptides}
  ...
}}
\source{
Source goes here
}
\usage{
biofilm
}
\description{
A description.
}
\keyword{datasets}
