% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_lfq.R
\name{convert_lfq}
\alias{convert_lfq}
\title{Title}
\usage{
convert_lfq(df, exd)
}
\arguments{
\item{df}{Description}

\item{exd}{Description}
}
\value{
Returns x
}
\description{
Title
}
\examples{
# example

}
