% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_val.R
\name{filter_val}
\alias{filter_val}
\title{Title}
\usage{
filter_val(data, op = c("==", "<=", ">="), pat, val)
}
\arguments{
\item{data}{Description}

\item{op}{Description}

\item{pat}{Description}

\item{val}{Description}
}
\value{
Returns x
}
\description{
Title
}
\examples{
# example

}
