% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nta.R
\docType{data}
\name{nta}
\alias{nta}
\title{Title}
\format{An object of class \code{data.frame} with 42,000 observations (rows) and 7 variables (columns).

Columns:

\describe{
  \item{\strong{particle_size}}{Description}
  \item{\strong{Strain}}{Description}
  \item{\strong{Bio_rep}}{Description}
  \item{\strong{Dilution}}{Description}
  \item{\strong{Tech_rep}}{Description}
  \item{\strong{Count}}{Description}
  \item{\strong{True_count}}{Description}
}}
\source{
Source goes here
}
\usage{
nta
}
\description{
A description.
}
\keyword{datasets}
