% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_go_bar.R
\name{plot_go_bar}
\alias{plot_go_bar}
\title{Visualise GO term coverage}
\usage{
plot_go_bar(go_table, terms_list, col_pal)
}
\arguments{
\item{go_table}{data.frame: output of \code{\link{process_fungifun}}}

\item{terms_list}{numeric vector: row numbers of the GO terms to be plotted}

\item{col_pal}{character vector (length = 3): names of three colours to use
for indicating BP, CC, MF.}
}
\value{
Returns a bar plot of the specified GO terms (generated by
\code{\link[ggplot2]{ggplot}}).
}
\description{
\code{plot_go_bar} generates a bar plot of specific GO terms from
the output of \code{\link{process_fungifun}}.

The length of the bars indicates the percentage of the all the GO term proteins
that are present in the protein list used to query FungiFun2. The colour of the
bars indicates the specific GO domain a GO term is under: biological process
(BP), cellular component (CC), molecular function (MF).
}
\examples{
\dontrun{
# import a FungiFun2 output .csv file
go_table <- process_fungifun("./vignettes/figures/f4_GO_DAY_Y.csv")

# define a colour palette
go_pal <- c("#6D2077", "#E4002B", "#FFB81C")

# generate a GO term bar plot
plot_go_bar(go_table = go_table,
            terms_list = c(1:4, 10),
            col_pal = go_pal) +
 annotate("text", x = 1, y = 50, label = "bold('DAY286 Y')", parse = TRUE) +
 labs(fill = "GO domain") +
 theme(legend.text = element_text(size = 10),
       legend.title = element_text(size = 10))
}

}
