% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_imputation2.R
\name{plot_imputation2}
\alias{plot_imputation2}
\title{Visualise effect of imputation on protein intensities}
\usage{
plot_imputation2(exd, mat, ...)
}
\arguments{
\item{exd}{data.frame: contains information for plot labels. See the
examples section of \code{\link{assert_exd}} or \code{\link{atcc_exp}} for
examples of appropriately structured experimental designs.}

\item{mat}{numeric matrix: LFQ intensity data before imputation}

\item{...}{numeric matrix/matrices: LFQ intensity data after imputation.
Multiple matrices can be input for plotting.}
}
\value{
Returns density plots of all input matrices, e.g. before and
after imputation (generated by \code{\link[ggplot2]{ggplot}}).
}
\description{
\code{plot_imputation2} generates density plots of protein
LFQ intensities showing the distribution of intensities
before and after imputation.

This function has been taken directly from the \code{plot_imputation}
function in the \href{http://doi.org/10.18129/B9.bioc.DEP}{DEP} package and
then modified for use in this package.

\href{http://doi.org/10.18129/B9.bioc.DEP}{DEP} is distributed with the
Artistic-2.0 license
which requires any source code modifications to be clearly stated.

\code{plot_imputation2} includes the following modifications
of \code{plot_imputation}:

\itemize{
\item Dependance on a "SummarizedExperiment" object as input has been
removed. Instead the input has been changed to a numeric matrix.
\item Use of \code{theme_DEP1()} has been replaced with \code{theme_bw()}
}
}
\examples{
# load dplyr
library(dplyr)

# load a proteinGroups data.frame supplied with this package
my_proteinGroups <- atcc

# load its corresponding experimental design
my_expDesign <- atcc_exp

# filter for proteins identified with minimum 3 unique peptides
# and convert to numeric matrix
my_lfq <- my_proteinGroups \%>\%
  filter(Unique.peptides >= 3) \%>\%
  convert_lfq(., my_expDesign)

# filter for proteins quantified in min 2/3 reps of
# at least 1 sample group
my_filt <- my_lfq \%>\%
  filter_na4(., logic = "or", op = "<=",
             pat1 = "A10231_EV", val1 = 1,
             pat2 = "A10231_W", val2 = 1,
             pat3 = "A90028_EV", val3 = 1,
             pat4 = "A90028_W", val4 = 1)

# normalise LFQ intensities
my_norm <- limma::normalizeCyclicLoess(my_filt)

# impute missing values
my_imp <- impute_QRILC(my_norm)

# impute missing values with another method
my_imp2 <- imputeLCMD::impute.MinProb(my_norm)

# density plot before and after imputations
plot_imputation2(exd = atcc_exp,
                 my_norm,
                 my_imp,
                 my_imp2)
}
