% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mds.R
\name{plot_mds}
\alias{plot_mds}
\title{Title}
\usage{
plot_mds(mat, mat_labels, shape_size = 2, ...)
}
\arguments{
\item{mat}{Description.}
}
\value{
Description.
}
\description{
Description.
}
\examples{
# example

}
