% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Title}
\usage{
plot_volcano(data, p_val, log2fc, group, label_fun, point_size = 1,
  label_size = 2, x_lim = c(-10, 10), y_lim = c(0, 10),
  legend_params = c("left", 11, 10), axis_params = c("black", 10))
}
\arguments{
\item{data}{Description.}

\item{p_val}{Description.}

\item{log2fc}{Description.}

\item{group}{Description.}

\item{label_fun}{Description.}

\item{point_size}{Description.}

\item{label_size}{Description.}

\item{x_lim}{Description.}

\item{y_lim}{Description.}

\item{legend_params}{Description.}

\item{axis_params}{Description.}
}
\value{
Description.
}
\description{
Description.
}
\examples{

# example

}
