% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Title}
\usage{
plot_volcano(data, p_val, lfc, grp, colour = c("red1", "blue1"), label)
}
\arguments{
\item{data}{Description.}

\item{p_val}{Description.}

\item{lfc}{Description.}

\item{grp}{Description.}

\item{colour}{Description.}

\item{label}{Description.}
}
\value{
Description.
}
\description{
Description.
}
\examples{
#example

}
