% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeast_exp.R
\docType{data}
\name{yeast_exp}
\alias{yeast_exp}
\title{Yeast LC-MS/MS experimental design}
\format{A data frame with x rows and y variables:
\describe{
  \item{ID}{Description}
  \item{label}{Description}
  \item{condition}{Description}
  \item{replicate}{Description}
}}
\source{
Made by hand
}
\usage{
yeast_exp
}
\description{
A description.
}
\keyword{datasets}
